+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/PartialDataHandler.java
@@ -39,7 +39,7 @@ public class PartialDataHandler {
      * present for a given combination of request metrics and dimensions (pulled from the API request and generated
      * druid query) at the specified granularity.
      *
-     * @param constraints  Constraint containing all the column names the request depends on
+     * @param constraint  Constraint containing all the column names the request depends on
      * @param physicalTables  the tables whose column availabilities are checked
      * @param requestedIntervals  The intervals that may not be fully satisfied
      * @param granularity  The granularity at which to find missing intervals
@@ -47,13 +47,13 @@ public class PartialDataHandler {
      * @return subintervals of the requested intervals with incomplete data
      */
     public SimplifiedIntervalList findMissingTimeGrainIntervals(
-            DataSourceConstraint constraints,
+            DataSourceConstraint constraint,
             Set<PhysicalTable> physicalTables,
             @NotNull SimplifiedIntervalList requestedIntervals,
             Granularity granularity
     ) {
         SimplifiedIntervalList availableIntervals = physicalTables.stream()
-                .map(table -> table.getAvailableIntervals(constraints))
+                .map(table -> table.getAvailableIntervals(constraint))
                 .flatMap(SimplifiedIntervalList::stream)
                 .collect(SimplifiedIntervalList.getCollector());