+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/config/table/BaseTableLoaderSpec.groovy
@@ -60,7 +60,7 @@ class BaseTableLoaderSpec extends Specification {
 
     def "table group has correct contents after being build"() {
         when:
-        TableGroup group = loader.buildTableGroup(
+        TableGroup group = loader.buildDimensionSpanningTableGroup(
                 apiNames,
                 metricNames,
                 physDefs,